Fix file system delete of folder
Deleting a folder without recursive enabled failed as the used
nodejs API unlink does not support folder deletion.
Fixing it by first checking if it is a folder, and if so
using rmdir instead.

Added a simple unit test, which also exposed that
the folder deletion also fails when using trash support,
but that is a different issue.

Signed-off-by: Daniel Friederich <d.friederich@samsung.com>