[license] Add vscode license, referred-to in copied source code
In various source files in the code base, code originally from the vscode
repository, was re-used. In these instances, there is a copyright header,
like below, that refers to the file containing the full license:

/*---------------------------------------------------------------------------------------------
 *  Copyright (c) Microsoft Corporation. All rights reserved.
 *  Licensed under the MIT License. See License.txt in the project root for license information.
 *--------------------------------------------------------------------------------------------*/

I think we missed adding the full license text, as is mandated by the MIT
license.

This commit adds it, renamed for clarity. I think it should still be easy to
find, and should be less likely to cause confusion about what license is the
main one, in this repository/project.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>