Make browserWindow of splashScreen transparent
Currently an image with a transparent background is rendered with a white background.
This looks weird and is not the expected behavior when a transparent image is used.
With this change the browserWindow is set to transparent, so that the background is not rendered.
If adoptors want a white background they can simply use an image with a white background (or any other color).