Fix performance issues in terminal
Do not search for matches in the workspace when creating links in the
terminal, as this impacts performance in large workspaces. Instead,
search for a match only when the user wants to open a link. If the match
can't be found, open the quick input service and let the user select
from a list of possible matches.

Additionally, the LocalFileLinkProvider no longer matches single words,
to avoid polluting the terminal with links.

Contributed on behalf of STMicroelectronics