feat: implement language-model-registry
Implements the LanguageModelProviderRegistry which is able to handle
an arbitrary number of LanguageModelProviders.

Refactors the LanguageModelProvider to only return a simple text or
stream of text. It's now the agent's responsibility to convert this
into response parts. Therefore the interfaces are also moved to the
agent package.

The LanguageModelProviderRegistry implementation for the frontend
handles all LanguageModelProvider registered in the frontend as well as
in the backend.

Fixes the StreamNode in the tree-widget to update itself correctly
when new tokens arrive.