feat: introduce chat service and chat models
Introduces ChatModel, including nested ChatRequestModel and
ChatResponseModel to represent chat sessions. The chat models allow to
inspect and track requests and their responses.

Also introduces the ChatService which can be used to manage chat
sessions and sending requests.

Architecture is inspired by the VS Code implementation, however it
intends to be more generic.