Fix window maximization when using splash screen
Avoid eager restore of maximized state in `TheiaElectronWindow.restoreMaximizeState`
Calling `window.maximize()` implicitly also makes the window visible.
Therefore we have to check if the window is already visibible before restoring t he maximizeState.
If the window is not yet visible yet we restore the state once the window is shown.

Fixes #14218