feat(ai): enable querying history by session
In more complex scenarios, there are often multiple agents involved, which may call each other. In those cases it is crucial to allow tracing all LLM requests for a session across agents.

This change introduces a query to obtain all communication events across a session, irrespectively of the agent.

Contributed on behalf of STMicroelectronics.