Update PR template
Removed the changelog requirement from the review guidelines.
Changelog entries will be created in the release process, with the exception of breaking changes.
For breaking changes the message needs to be added within the PR.
Adjusted the PR template with sections to indicate if breaking changes are present.
Added hint that if there are breaking changes a changelog message is needed.
Streamlined the pull request guide to link directly to the existing PR template.
Added attribution field to PR template to make it easier to see during the release.