Address review comments
Implement a brief hysteresis also on the dismissal of the hover pop-up
on mouse movement out of the target element so that

(a) the user has an opportunity to mouse over the pop-up itself to
    interact with it, clicking links, hitting buttons, copying text
(b) not dismiss the pop-up until the mouse pointer stops moving (helps
    accessibility for those with pointer accuracy challenges)

The previous commit also had a misunderstanding of the the disposable
returned by the DisposableCollection::push() method, that it would also
dispose the original disposable, but it doesn't. Reworking disposable
management fixed other issues in which transitioning to another hover
target sometimes would not pop up.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>