chore(ai): enable chat input on chat widget activate
We get warnings in the console from the chat widget as it doesn't accept focus. More importantly, it is cumbersome having to click the input field to start typing.

This change therefore focuses the input field when the chat widget is activated. Also it shows the placeholder text until a text is entered, which is a nicer touch than before.

Contributed on behalf of STMicroelectronics.