Cancel pending hover on mouse exit
If the mouse exits before the HoverService has rendered its pop-up,
cancel the pending presentation of that pop-up.

Also fix the problem of it being difficult to mouse over to the pop-up
itself after it is shown to interact with it (it being necessary to
carefully track over the bottom triangle). This is accomplished by
cancelling only after a short delay that allows the mouse to reach
either the pop-up or the original hover target.

Fixes #14532

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>