Fix updating breakpoints when debugging starts
Previously, when starting to debug, the breakpoints did not
move to valid lines. The breakpoints were send to the debug
adapter before configurationDone and some debug adapters
(e. g. python) return updated positions directly. The updates
were not visible in the editor, because the connection was not
marked as initialized yet and events were ignored. The
updates became only visible, when changing breakpoints later.
This commit fixes this.

Fixes #14096

Signed-off-by: Florian Richter <florian.richter@mvtec.com>