fix: add timeout for early evaluate DAP requests
The `vscode-js-debug` does not reject when the evaluate request arrives
too early to the debug adapter and the evaluate handler is not yet
registered.

Ref: https://github.com/microsoft/vscode-js-debug/blob/d7730405d64e81eb364f76ea38efb7c93a5f877d/src/dap/connection.ts#L172

If the first evaluate request never rejects nor resolves, none of the
following evaluate requests will as they are queued, making the Theia
UI never update.

This commit adds a 1000ms timeout to the evaluate DAP call.

Note that the VS Code Debug JS extension cannot accept the evaluate
request despite it already sent the initialized to the client.

Closes eclipse-theia/theia#11955