Add prompt template AI variable
Adds a new AI variable that resolves registered prompt templates based on their ID.
The variable resolves to the resolved prompt template. I.e. variables and functions in the template are already resolved in the returned value.

In the chat a prompt template with ID myprompt is referenced as #prompt:myprompt

Adds prompt template id autocompletion for the chat window that triggers on entering the variable separator `:`