Option to put close icons at the start of the tab label on macOS
Add a new window.tabCloseIconPlacement preference for whether to present
the Close (X) icon in tab titles on the left or the right of the tab
in conformity with the OS's native tab controls.
Render the tab title with the icon on the left or right accordingly.
Show the new preference in the Settings UI on macOS platform only.
Default to the right as on other platforms to maintain current appearance.

Fixes eclipse-theia/theia-ide#460

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>