Put close icons at the start of the tab label by default on macOS
Add a new window.tabCloseIconPlacement preference for whether to present
the Close (X) icon in tab titles on the left or the right of the tab.
Default to the left on macOS platform in conformity with the OS's native
tab controls.
Show the new preference in the Settings UI on macOS platform only.

Render the tab title with the icon on the left or right accordingly.

Fixes eclipse-theia/theia-ide#460

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>