chore: use shared packages in src-gen
Src-gen is executed in the environment controlled by adopters. In the
generated code dependencies are imported. If these dependencies are not
located where webpack expects them, the adopter will experience build
errors. This is not a common case, but it does happen in practice.

The generated code is made more robust by reusing the shared core
dependencies where applicable. As it's common to declare '@theia/core'
as a dependency within the adopter package, this more commonly results
in a working build.

Resolves #15173

Contributed on behalf of STMicroelectronics