fix: handle contextmenu events on macOS (#15208)
Work around failure of Lumino to handle contextmenu events on macOS

In the contextmenu event handler of the Theia tab bar, do not quash the
event as the superclass does. Also acknowledge that on macOS platform
there are mouse events that summon the context menu and let those
bubble up instead of trying to handle them as other kinds of mouse
pointer input.

Fixes #15108

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>