ai: Support prompt fragments and recursive variable resolving (#15196)
fixes #14899
fixes #15000

* Add prompt template AI variable

Adds a new AI variable that resolves registered prompt templates based on their ID.
The variable resolves to the resolved prompt template. I.e. variables and functions in the template are already resolved in the returned value.

In the chat a prompt template with ID myprompt is referenced as #prompt:myprompt

Adds prompt template id autocompletion for the chat window that triggers on entering the variable separator `:`
and automatically after the `#prompt` variable has been autocompleted.

* Resolve functions after variables in prompt service

Functions must not be immediately resolved in prompt fragments because function objects are lost this way.
Instead, they are left in and resolved when the final chat message or prompt containing a prompt fragment is resolved.
For prompts, all variables must be resolved before resolving functions to allow resolving functions in resolved variables.

- Resolve functions after variables when getting a prompt
- Extend PromptService with getPromptFragment method that resolves variables but not functions
- ChatRequestParser can now handle chat and prompt function formats
- Add unit test for prompt service testing resolving a function within a variable

* Resolve functions in resolved variables in chat request parser

- Move variable resolvement from chat service to chat request parser
- Resolve functions in resolved variable texts
- Add unit test for this

* Support explicit AI variable dependencies and recursive resolvement

- Add interface AIVariableResolverWithVariableDependencies for variable resolvers that want to resolve dependencies.
  The variable service hands in a resolve method to these resolvers.
- The AI variable service now recursively resolves all variables while providing cycle detection.
  If a cycle is found, the resolvement of only the recursive branch is stopped.