fix: autocomplete content in chat view (#15240)
Fixes autocomplete content not showing up in inline editors like the
ChatInput.

With the Lumino migration we disabled 'fixedOverflowWidgets' in our
Monaco editors, as Lumino widget styles leverages 'contain',
influencing the position of autocomplete content.

This is now reverted, and instead the 'contain' of Lumino widgets is
forcefully disabled, fixing the autocomplete issue.

fixes #15237

Co-authored-by: Eugen Neufeld <eneufeld@eclipsesource.com>