Fix autostart stopping for following servers if one fails
Before this the autostart would stop for all following servers if one failed.
This was due to uncatched errors when trying to resolve the tools.
To prevent this in the future, improved the handling of the getTools() on client side.
Also adapted the Start Command to use the new status reporting instead of calling getTools() again.