Set selection upon editor navigation after editor is fully visible
This change required the introduction of "navigation transactions",
where a location is not remembered in the navigation location stack
while another navigation transaction is still active. This prevents
the navigation location service from recording intermediate locations
while navigating to an already open editor. Only the final location in
simultaneous transactions is recorded.

Fixes #15145, #15191

Contributed on behalf of STMicroelectronics

Signed-off-by: Thomas Mäder <t.s.maeder@gmail.com>