feat(chat): introduce LLM-based chat naming and last interaction date (#15116)
Finding chat sessions is currently very difficult as we are recoding the last request as a name. It'd be better to use a language model to generate a name for the chat session and also record the last interaction date, to have proper sorting and a chronological relation to the chat sessions.

Moreover, this PR allows users to manually specify a name for the chat session, which will be used instead of the generated name.

Fixes https://github.com/eclipse-theia/theia/issues/15115

Co-authored-by: Stefan Dirix <sdirix@eclipsesource.com>