Do not re-export the configuration folder as env variable
Before this the configurationFolder (which includes the user settings for example) was exported as a env variable THEIA_CONFIG_DIR.
This is problematic, as we also allow users to define a configurationFolder via this variable,
and use it whenever provided for a new session.
Therefore, if you debug another application from a Theia tool (e.g. the Theia IDE) you will always get the tools configurationFolder, which is probably not intended.
As this results in settings being changed in the started application to also be applied in the root too (e.g. colorTheme).

Signed-off-by: Simon Graband <sgraband@eclipsesource.com>