Adds prompt fragments configuration widget
Introduces a new widget for configuring AI prompt fragments and system prompts.
Displays all system prompts, which agent they are used by, their variants and customizations.
Shows the active and default variants as well as the active customization.
Offers buttons to revert to a certain customization or deleting it.
Allows the user to create a customization of built-in prompt fragents.
Displays the template text for each customization.
Shows prompt fragments that are not associated with and agent at the end.

Signed-off-by: Simon Graband <sgraband@eclipsesource.com>