fix: check for a valid LanguageModelStreamResponsePart
OpenAI returns empty stream parts which create an empty
newContent object and is added as a markdown part in the section not expecting a text part. This breaks all the logic.

A new check is added that checks for known stream parts.
Unknown parts are logged as a warning.

fixes #15646