Prompt management refactoring and view (#15632)
Unified naming:
- Introduced clear separation between prompt templates (text) and prompt fragments (template + metadata)
- Added support for 0...1 built-in versions per fragment and unlimited custom versions
- Agents are now assigned prompts instead of prompt templates
- Agent prompts are a set of variant (1...n), which are prompt fragments, united under one id.
- Added selected & active variant logic for system prompts

The prompt service is now responsible for handling all requests regarding prompts.
This includes customization's (will be forwarded to the customization service, if one exists).
Also handles the newly introduced agent prompts.

Changed agents prompts declaration:
- Agent now provide the id of the system prompt they are using.
- Additionally they need to define the default variant, but can register many more variants.

Improved logic in prompt-service:
- Added listener when the prompts have changed
- Removed as much assumed information about the customization service as possible (e.g. file name)
- Added PromptVariantSets, selected & active variant maps to keep the state.
- Introduces types for BuiltInPromptFragment & CustomizedPromptFragment.

Introduces a new widget for configuring AI prompt fragments and PromptVariantSets.
Displays all PromptVariantSets, which agent they are used by, their variants and customizations.
Shows the active and selected variants as well as the active customization.
Offers buttons to revert to a certain customization or deleting it.
Allows the user to create a customization of built-in prompt fragents.
Displays the template text for each customization.
Shows prompt fragments that are not associated part of a variant set.