feat: implement prompt caching for Anthropic
This is a backport of #15731 for Theia 1.61.x

Mark system prompts, tools and messages for caching in the Anthropic
LanguageModel. For typical agent workflows like "@Coder" this reduces
token costs by a large factor and reduces the pressure on the rate
limit.

Currently caching is enabled by default for all Anthropic models.

Also updates token tracking functionality to be aware of token caching
and updates the Anthropic SDK.