fix: remove obsolete menu workaround
We previously used a workaround in the browser-menu-plugin to support
initially empty menus, which weren’t compatible with Lumino.

Since the menu refactoring in #14676, this workaround is no longer
needed. In fact, it now causes issues like empty submenus appearing in
context menus. Removing the workaround fixes this regression.

Fixes #15694