Do not attempt to update actions in non-existent menus
The MenuModelRegistry now throws on attempts to request the menu for an
ID that is not registered. As at least one of the menus that the
DirtyDiffWidget does not exist in typical installations of Theia IDE,
this was causing the diffs not to appear. Because the widget doesn't
need to update actions in menus that don't exist, it takes a more
lenient approach.

Fixes eclipsesource/theia/issues#253

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>