feat: theia-ai agent completion notification
To be compliant with existing behavior, the default notification setting
is 'off'. The global default can be set in the preference
ai-features.notifications.default.

The preference ai-features.agentSettings.<agent-id>.completionNotification
can be used to override the global notification setting for a specific
theia-ai agent.

Notifications will only be shown if the Chat Widget does not have the
focus.

The following notification settings are supported:
- 'off': no notification.
- 'os-notification': Displays a notification in the operating system or
  the web browser, depending on whether electron is used or not
- 'message': uses the theia-message service
- 'blink': blinks the titlebar + show window if it is hidden

Fixes eclipsesource/theia#223

Signed-off-by: Olaf Lessenich <olessenich@eclipsesource.com>
Co-authored-by: Johannes Faltermeier <jfaltermeier@eclipsesource.com>
Co-authored-by: Eugen Neufeld <eneufeld@eclipsesource.com>
Co-authored-by: Simon Graband <sgraband@eclipsesource.com>