feat: enable specifying the original state of change set files
I'm working on a use case where I have an external agent that I integrate as a Theia AI agent. This external agent modifies workspaces resources directly and only notifies our Theia agent after the change, but it'll store a backup of the original in the file system, which we can obtain.

I want to use this in order to track and show file changes in a change set, so the user has a central place where they see all applied changes. Therefore, I experimented with extending the change set file element with an `originalState` that is passed from the agent, instead of being read from the file system.