Add AI model aliases and default models (#15782)
Initial implementation of #15395

Introduces support for AI model aliases to facilitate updating many agents' used models at once.
An alias either points to a single model selected by the user or a chain of default/fallback models to consider in priority order.
For the fallback chain, the first ready model (e.g. has an API key configured) is used. For this the language model was extended with a status to be set by its manager.

- Introduces a language model alias registry for registering and retrieving model aliases and chains
- Adds support for chains of default model aliases (`code`, `universal`, `code completion`, `summarize`)
- Selects the first available model (with an API key) from the chain when resolving an alias; falls back to the first in the chain if none are ready
- Enables users to point an alias to use a specific model
- Extends language models to track readiness (e.g., whether an API key is set)
- Adds a new Model Aliases tab to the the AI configuration UI to:
  - Display alias chains
  - Select a mode for an alias

Signed-off-by: Simon Graband <sgraband@eclipsesource.com>
Co-authored-by: Simon Graband <sgraband@eclipsesource.com>
Co-authored-by: Stefan Dirix <sdirix@eclipsesource.com>