feat(chat): cancel incomplete requests on new request
If a user sends a new request to a session that currently as an incomplete request, we should always cancel those first before adding a new request. Otherwise the state becomes very unclear (what was the context of the new request while the past ones are still generating).

This also makes it more intuitive when submitting prompts while a previous request is waiting for input or generating, as we show the cancel button in the input.