fix: add argument processor for plugin toolbar command arguments (#16230)
If commands are triggered from plugin-contributed toolbar items (main or widget toolbars) and the widget is passed as an argument, process the argument and return only the widget ID. This prevents serialization errors and allows the command to execute properly.

This not only fixes custom toolbar items that add extension-contributed commands, it also fixes widget toolbar actions in extension views (if a corresponding handler is registered).

Part of GH-15932
Resolves GH-16223