dev: prevent await identifiers
Adjusts the eslint rules to prevent shadowing restricted names. Also
explicitly forbids using the name "await" as an identifier because it is
not covered by the built-in rule.

This avoids issues with adopters using esbuild. The esbuild parser is
overly strict and throws errors in these cases, although they are valid
Javascript.

Contributed on behalf of STMicroelectronics