[vscode] chore: align module declarations for proposed.*.d.ts
We used mixed declarations of the proposed plugin modules, this commit aligns them to use export consistently.
As they are imported in the theia.d.ts which is the main definition of @theia/plugin, we do not want to re declare them in each proposed type file.
The typedoc documentation benefits from that as well, this way the @theia/plugin types are correctly listed under the theia.d.ts entry then.

Contributed on behalf of STMicroelectronics