feat: Terminal manager for multiple terminals in one view (#16604)
Introduces a new extension terminal-manager providing a terminal manager widget that enables managing multiple terminals inside a single view.
It supports splitting vertically and horizontally and supports multiple pages.

The extension provides a setting terminal.grouping.mode to configure whether new terminals are created as separate views (as before) or inside the terminal manager. For now, the old behavior is the default.

The terminal-manager extension is added to the browser and electron example applications.

---------

Co-authored-by: eeakrnm <kenneth.marut@ericsson.com>
Co-authored-by: Colin Grant <cgrant.ext@eclipsesource.com>