fix: duplicate command check
On startup we can observe warnings like:
```
logger-protocol.ts:118 2025-12-05T18:17:54.751Z root WARN Command name 'clear' is used by multiple fragments: claude-code-slash-clear and claude-code-slash-clear
```

This is caused by a check for duplicates AFTER the command was actually already added. Also it does not take the agents for which it was defined into account.